mqtt-iot-router/requirements.txt
#
####### How to install requirements #######
#
# pip install -r requirements.txt
#
###########################################
#

paho-mqtt
json-cfg
